kafka: improve consumer test stability
This commit improves the stability of the consumer test by reducing the
`FetchMaxWait` to 100ms.
