+++ kafka/producer_test.go
@@ -32,7 +32,6 @@ import (
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/twmb/franz-go/pkg/kfake"
 	"github.com/twmb/franz-go/pkg/kgo"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/sdk/trace/tracetest"
@@ -312,26 +311,8 @@ func TestProducerConcurrentClose(t *testing.T) {
 	wg.Wait()
 }
 
-func newClusterAddrWithTopics(t testing.TB, partitions int32, topics ...string) []string {
-	t.Helper()
-	cluster, err := kfake.NewCluster(kfake.SeedTopics(partitions, topics...))
-	require.NoError(t, err)
-	t.Cleanup(cluster.Close)
-
-	return cluster.ListenAddrs()
-}
-
-func newClusterWithTopics(t testing.TB, partitions int32, topics ...string) (*kgo.Client, []string) {
-	t.Helper()
-	addrs := newClusterAddrWithTopics(t, partitions, topics...)
-
-	client, err := kgo.NewClient(kgo.SeedBrokers(addrs...))
-	require.NoError(t, err)
-
-	return client, addrs
-}
-
 func newProducer(t testing.TB, cfg ProducerConfig) *Producer {
+	t.Helper()
 	producer, err := NewProducer(cfg)
 	require.NoError(t, err)
 	t.Cleanup(func() {
